moved the Bzorg downloader to a seperate module
