+++ stormstats/__init__.py
@@ -1,1 +1,2 @@
 from .storm import *
+from .downloader import *
